get rid of the ApiRouteServiceProvider and the ApiBaseRouteServiceProvider
